Move `aniblastall`-specific tests to `test_aniblastall.py`
