+++ tests/test_anib.py
@@ -119,65 +119,6 @@ def anib_output_dir(dir_anib_in):
     )
 
 
-# Test legacy BLAST (blastall) command generation
-def test_blastall_dbjobdict(path_fna_all, tmp_path):
-    """Generate dictionary of legacy BLASTN database jobs."""
-    blastcmds = anib.make_blastcmd_builder("ANIblastall", tmp_path)
-    jobdict = anib.build_db_jobs(path_fna_all, blastcmds)
-    expected = [
-        (tmp_path / _.name, f"formatdb -p F -i {tmp_path / _.name} -t {_.stem}")
-        for _ in path_fna_all
-    ]
-    assert sorted([(k, v.script) for (k, v) in jobdict.items()]) == sorted(expected)
-
-
-def test_blastall_graph(path_fna_all, tmp_path, fragment_length):
-    """Create jobgraph for legacy BLASTN jobs."""
-    fragresult = anib.fragment_fasta_files(path_fna_all, tmp_path, fragment_length)
-    blastcmds = anib.make_blastcmd_builder("ANIblastall", tmp_path)
-    jobgraph = anib.make_job_graph(path_fna_all, fragresult[0], blastcmds)
-    # We check that the main script job is a blastn job, and that there
-    # is a single dependency, which is a makeblastdb job
-    for job in jobgraph:
-        assert job.script.startswith("blastall -p blastn")
-        assert len(job.dependencies) == 1
-        assert job.dependencies[0].script.startswith("formatdb")
-
-
-def test_blastall_multiple(path_fna_two, tmp_path):
-    """Generate legacy BLASTN commands."""
-    cmds = anib.generate_blastn_commands(path_fna_two, tmp_path, mode="ANIblastall")
-    expected = [
-        (
-            "blastall -p blastn -o "
-            f"{tmp_path / str(path_fna_two[0].stem + '_vs_' + path_fna_two[1].stem + '.blast_tab')} "
-            f"-i {path_fna_two[0]} "
-            f"-d {path_fna_two[1]} "
-            "-X 150 -q -1 -F F -e 1e-15 -b 1 -v 1 -m 8"
-        ),
-        (
-            "blastall -p blastn -o "
-            f"{tmp_path / str(path_fna_two[1].stem + '_vs_' + path_fna_two[0].stem + '.blast_tab')} "
-            f"-i {path_fna_two[1]} "
-            f"-d {path_fna_two[0]} "
-            "-X 150 -q -1 -F F -e 1e-15 -b 1 -v 1 -m 8"
-        ),
-    ]
-    assert cmds == expected
-
-
-def test_blastall_single(path_fna_two, tmp_path):
-    """Generate legacy BLASTN command-line."""
-    cmd = anib.construct_blastall_cmdline(path_fna_two[0], path_fna_two[1], tmp_path)
-    expected = (
-        f"blastall -p blastn -o {tmp_path / str(path_fna_two[0].stem + '_vs_' + path_fna_two[1].stem + '.blast_tab')} "
-        f"-i {path_fna_two[0]} "
-        f"-d {path_fna_two[1]} "
-        "-X 150 -q -1 -F F -e 1e-15 -b 1 -v 1 -m 8"
-    )
-    assert cmd == expected
-
-
 # Test BLAST+ (blastn) command generation
 def test_blastn_dbjobdict(path_fna_all, tmp_path):
     """Generate dictionary of BLASTN+ database jobs."""
@@ -248,30 +189,6 @@ def test_blastn_single(path_fna_two, tmp_path):
     assert cmd == expected
 
 
-# Test legacy BLAST database formatting (formatdb) command generation
-def test_formatdb_multiple(path_fna_two, tmp_path):
-    """Generate legacy BLAST db creation commands."""
-    cmds = anib.generate_blastdb_commands(path_fna_two, tmp_path, mode="ANIblastall")
-    expected = [
-        (
-            f"formatdb -p F -i {tmp_path / path_fna_two[0].name} -t {path_fna_two[0].stem}",
-            tmp_path / path_fna_two[0].name,
-        ),
-        (
-            f"formatdb -p F -i {tmp_path / path_fna_two[1].name} -t {path_fna_two[1].stem}",
-            tmp_path / path_fna_two[1].name,
-        ),
-    ]
-    assert cmds == expected
-
-
-def test_formatdb_single(path_fna, tmp_path):
-    """Generate legacy BLAST formatdb command-line."""
-    cmd = anib.construct_formatdb_cmd(path_fna, tmp_path)
-    expected = f"formatdb -p F -i {tmp_path / path_fna.name} -t {path_fna.stem}"
-    assert cmd[0] == expected
-
-
 # Test FASTA file fragmentation for ANIb methods
 def test_fragment_files(path_fna_all, tmp_path, dir_tgt_fragments, fragment_length):
     """Fragment files for ANIb/ANIblastall."""
@@ -317,26 +234,11 @@ def test_makeblastdb_single(path_fna, tmp_path):
 
 
 # Test output file parsing for ANIb methods
-def test_parse_legacy_blastdir(anib_output_dir):
-    """Parses directory of legacy BLAST output."""
-    orglengths = pyani_files.get_sequence_lengths(anib_output_dir.infiles)
-    fraglengths = anib.get_fraglength_dict(anib_output_dir.fragfiles)
-    result = anib.process_blast(
-        anib_output_dir.legacyblastdir, orglengths, fraglengths, mode="ANIblastall"
-    )
-    assert_frame_equal(
-        result.percentage_identity.sort_index(1).sort_index(),
-        anib_output_dir.legacyblastresult.sort_index(1).sort_index(),
-    )
-
-
 def test_parse_blastdir(anib_output_dir):
     """Parse directory of BLAST+ output."""
     orglengths = pyani_files.get_sequence_lengths(anib_output_dir.infiles)
     fraglengths = anib.get_fraglength_dict(anib_output_dir.fragfiles)
-    result = anib.process_blast(
-        anib_output_dir.blastdir, orglengths, fraglengths, mode="ANIb"
-    )
+    result = anib.process_blast(anib_output_dir.blastdir, orglengths, fraglengths)
     assert_frame_equal(
         result.percentage_identity.sort_index(1).sort_index(),
         anib_output_dir.blastresult.sort_index(1).sort_index(),
@@ -348,14 +250,3 @@ def test_parse_blasttab(anib_output):
     fragdata = anib.get_fraglength_dict([anib_output.fragfile])
     result = anib.parse_blast_tab(anib_output.tabfile, fragdata, mode="ANIb")
     assert (a == b for a, b in zip(result, [4_016_551, 93, 99.997_693_577_050_029]))
-
-
-def test_parse_legacy_blasttab(anib_output):
-    """Parses ANIB legacy .blast_tab output."""
-    fragdata = anib.get_fraglength_dict([anib_output.fragfile])
-    result = anib.parse_blast_tab(
-        anib_output.legacytabfile, fragdata, mode="ANIblastall"
-    )
-    assert (
-        a == b for a, b in zip(result, [1_966_922, 406_104, 78.578_978_313_253_018])
-    )
